Split init-subcommand logic out of cmd/mdsmith/main.go
The 2026-07-19 architecture audit found that main.go crossed the
project's documented ~1000-line threshold, almost entirely from the
new `mdsmith init --add` pack-scaffolding machinery. Move that logic
into cmd/mdsmith/init.go, matching the existing per-subcommand file
pattern (check.go, fix.go, deps.go, ...), with tests moved to a
matching init_unit_test.go. Pure move, no behavior change.

Also records the audit findings, files plans for the two lower-priority
tax findings (missing printInitCatalog unit test, duplicated isClaimed
helper), and splits architecture-audit-archive.md into a second shard
now that the first exceeded the file-length budget.
