+++ cmd/mdsmith/main_unit_test.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"gopkg.in/yaml.v3"
 
 	"github.com/jeduden/mdsmith/internal/bytelimit"
 	"github.com/jeduden/mdsmith/internal/config"
@@ -20,7 +19,6 @@ import (
 	fixpkg "github.com/jeduden/mdsmith/internal/fix"
 	"github.com/jeduden/mdsmith/internal/lint"
 	vlog "github.com/jeduden/mdsmith/internal/log"
-	"github.com/jeduden/mdsmith/internal/pack"
 	"github.com/jeduden/mdsmith/internal/query"
 	ruledocs "github.com/jeduden/mdsmith/internal/rules"
 )
@@ -1146,495 +1144,6 @@ func TestLoadConfigRaw_EmptyPath_ReturnsNonNilConfig(t *testing.T) {
 	assert.NotNil(t, cfg)
 }
 
-// --- runInit ---
-
-func TestRunInit_ExtraArgs_ExitsTwo(t *testing.T) {
-	captureStderr(func() {
-		code := runInit([]string{"extra"})
-		assert.Equal(t, 2, code)
-	})
-}
-
-func TestRunInit_CreatesConfigFile(t *testing.T) {
-	dir := t.TempDir()
-	oldWd, err := os.Getwd()
-	require.NoError(t, err)
-	defer func() { _ = os.Chdir(oldWd) }()
-	require.NoError(t, os.Chdir(dir))
-
-	captureStderr(func() {
-		code := runInit(nil)
-		assert.Equal(t, 0, code)
-	})
-
-	data, err := os.ReadFile(filepath.Join(dir, ".mdsmith.yml"))
-	require.NoError(t, err)
-	// Verify it's parseable YAML
-	var out map[string]any
-	require.NoError(t, yaml.Unmarshal(data, &out))
-}
-
-func TestRunInit_AlreadyExists_SkipsAndExitsZero(t *testing.T) {
-	dir := t.TempDir()
-	oldWd, err := os.Getwd()
-	require.NoError(t, err)
-	defer func() { _ = os.Chdir(oldWd) }()
-	require.NoError(t, os.Chdir(dir))
-	require.NoError(t, os.WriteFile(".mdsmith.yml", []byte("rules: {}\n"), 0644))
-
-	// A bare re-run over an existing config is idempotent: it leaves the
-	// config unchanged, notes it, and exits 0 (use --force to overwrite).
-	var code int
-	out := captureStderr(func() { code = runInit(nil) })
-	assert.Equal(t, 0, code)
-	assert.Contains(t, out, "already exists")
-	assert.Contains(t, out, "--force")
-	cfg, err := os.ReadFile(".mdsmith.yml")
-	require.NoError(t, err)
-	assert.Equal(t, "rules: {}\n", string(cfg), "existing config left unchanged")
-}
-
-func TestRunInit_Add_Wordlists_ScaffoldsCuratedFiles(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-
-	captureStderr(func() {
-		code := runInit([]string{"--add", "wordlists"})
-		assert.Equal(t, 0, code)
-	})
-
-	// A fresh project: init writes the default config and --add wordlists
-	// lands every curated list on disk. (Content is covered in the pack
-	// package's TestWordlistsPack_Files.)
-	_, err := os.Stat(filepath.Join(dir, ".mdsmith.yml"))
-	require.NoError(t, err, "default config written alongside the pack")
-	for _, name := range []string{"ai-speak", "ai-openers"} {
-		_, err := os.Stat(filepath.Join(dir, ".mdsmith", "wordlists", name+".yaml"))
-		assert.NoErrorf(t, err, "%s.yaml scaffolded", name)
-	}
-}
-
-func TestRunInit_Add_ExistingConfigScaffoldsAnyway(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// Already-initialized project: .mdsmith.yml exists. --add must still
-	// scaffold and leave the config unchanged (exit 0, no --force needed).
-	require.NoError(t, os.WriteFile(".mdsmith.yml", []byte("rules: {}\n"), 0o644))
-
-	captureStderr(func() {
-		code := runInit([]string{"--add", "wordlists"})
-		assert.Equal(t, 0, code, "--add must work on an initialized project")
-	})
-
-	cfg, err := os.ReadFile(".mdsmith.yml")
-	require.NoError(t, err)
-	assert.Equal(t, "rules: {}\n", string(cfg), "existing config left unchanged")
-	for _, name := range []string{"ai-speak", "ai-openers"} {
-		_, err := os.Stat(filepath.Join(dir, ".mdsmith", "wordlists", name+".yaml"))
-		assert.NoErrorf(t, err, "%s.yaml scaffolded", name)
-	}
-}
-
-func TestRunInit_Starter_ExistingConfigSkips(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	require.NoError(t, os.WriteFile(".mdsmith.yml", []byte("rules: {}\n"), 0o644))
-
-	// A config source over an existing config no longer errors: the config
-	// is left unchanged with a notice and the run exits 0.
-	var code int
-	out := captureStderr(func() { code = runInit([]string{"--starter", "okf"}) })
-	assert.Equal(t, 0, code, "--starter over an existing config skips, not errors")
-	assert.Contains(t, out, "already exists")
-	assert.Contains(t, out, "--force")
-	cfg, err := os.ReadFile(".mdsmith.yml")
-	require.NoError(t, err)
-	assert.Equal(t, "rules: {}\n", string(cfg), "existing config left unchanged")
-}
-
-func TestRunInit_Force_Overwrites(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	require.NoError(t, os.WriteFile(".mdsmith.yml", []byte("rules: {}\n"), 0o644))
-
-	captureStderr(func() {
-		code := runInit([]string{"--starter", "okf", "--force"})
-		assert.Equal(t, 0, code)
-	})
-	cfg, err := os.ReadFile(".mdsmith.yml")
-	require.NoError(t, err)
-	assert.Contains(t, string(cfg), "required-frontmatter", "--force overwrote with the okf starter")
-}
-
-func TestRunInit_StarterWithAdd_ExistingConfigSkipsConfigButScaffolds(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	require.NoError(t, os.WriteFile(".mdsmith.yml", []byte("rules: {}\n"), 0o644))
-
-	// The config source is skipped over the existing file, but the additive
-	// pack still applies — the two axes are independent, so a combined run
-	// no longer aborts the way it once did.
-	var code int
-	captureStderr(func() { code = runInit([]string{"--starter", "okf", "--add", "wordlists"}) })
-	assert.Equal(t, 0, code)
-	cfg, err := os.ReadFile(".mdsmith.yml")
-	require.NoError(t, err)
-	assert.Equal(t, "rules: {}\n", string(cfg), "existing config left unchanged")
-	for _, name := range []string{"ai-speak", "ai-openers"} {
-		_, err := os.Stat(filepath.Join(dir, ".mdsmith", "wordlists", name+".yaml"))
-		assert.NoErrorf(t, err, "%s scaffolded despite the config skip", name)
-	}
-}
-
-func TestRunInit_FromMarkdownlintWithAdd_ExistingConfigSkipsConfigButScaffolds(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	require.NoError(t, os.WriteFile(".mdsmith.yml", []byte("rules: {}\n"), 0o644))
-
-	// The other config source behaves identically: the existing config is
-	// skipped (no markdownlint discovery even runs) while --add scaffolds.
-	var code int
-	captureStderr(func() { code = runInit([]string{"--from-markdownlint", "--add", "wordlists"}) })
-	assert.Equal(t, 0, code)
-	cfg, err := os.ReadFile(".mdsmith.yml")
-	require.NoError(t, err)
-	assert.Equal(t, "rules: {}\n", string(cfg), "existing config left unchanged")
-	for _, name := range []string{"ai-speak", "ai-openers"} {
-		_, err := os.Stat(filepath.Join(dir, ".mdsmith", "wordlists", name+".yaml"))
-		assert.NoErrorf(t, err, "%s scaffolded despite the config skip", name)
-	}
-}
-
-func TestRunInit_UnknownFlag_ExitsTwo(t *testing.T) {
-	var code int
-	out := captureStderr(func() { code = runInit([]string{"--definitely-not-a-flag"}) })
-	assert.Equal(t, 2, code)
-	assert.Contains(t, out, "unknown flag")
-}
-
-func TestRunInit_ConfigPathIsDir_WriteError(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// .mdsmith.yml is a directory, which statTarget reports as absent, so
-	// init proceeds to write and os.WriteFile fails with "is a directory" —
-	// exercising writeInitConfig's write-error branch.
-	require.NoError(t, os.Mkdir(".mdsmith.yml", 0o755))
-
-	var code int
-	out := captureStderr(func() { code = runInit([]string{"--force"}) })
-	assert.Equal(t, 2, code)
-	assert.Contains(t, out, "writing .mdsmith.yml")
-}
-
-func TestRunInit_SymlinkConfig_Refused(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// A symlinked .mdsmith.yml must not be written through, even with
-	// --force: init refuses instead of following the link out of the tree.
-	require.NoError(t, os.Symlink(filepath.Join(dir, "target.yml"), ".mdsmith.yml"))
-
-	var code int
-	out := captureStderr(func() { code = runInit([]string{"--force"}) })
-	assert.Equal(t, 2, code)
-	assert.Contains(t, out, "symlink")
-	_, statErr := os.Stat(filepath.Join(dir, "target.yml"))
-	assert.True(t, os.IsNotExist(statErr), "must not write through the symlink")
-}
-
-func TestRunInit_Add_TrimsWhitespace(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// pflag splits "wordlists, " on the comma into ["wordlists", " "]; the
-	// blank entry and the surrounding space must be normalized away so the
-	// pack still resolves instead of failing as an unknown name.
-	var code int
-	captureStderr(func() { code = runInit([]string{"--add", "wordlists, "}) })
-	assert.Equal(t, 0, code)
-	_, err := os.Stat(filepath.Join(dir, ".mdsmith", "wordlists", "ai-speak.yaml"))
-	require.NoError(t, err, "trimmed pack name still scaffolds")
-}
-
-func TestRunInit_Add_UnknownPack_ExitsTwo(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-
-	var code int
-	out := captureStderr(func() { code = runInit([]string{"--add", "bogus"}) })
-	assert.Equal(t, 2, code)
-	assert.Contains(t, out, `unknown pack "bogus"`)
-	// Fails before writing anything: an invalid pack name never touches the
-	// config or the .mdsmith directory.
-	_, err := os.Stat(filepath.Join(dir, ".mdsmith.yml"))
-	assert.True(t, os.IsNotExist(err), "no config written when a pack name is invalid")
-}
-
-func TestRunInit_Add_ScaffoldError(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// .mdsmith is a regular file, so the pack's parent-directory checks
-	// fail and runInit must surface the scaffold error as exit 2.
-	require.NoError(t, os.WriteFile(".mdsmith", []byte("x"), 0o644))
-
-	captureStderr(func() {
-		code := runInit([]string{"--add", "wordlists"})
-		assert.Equal(t, 2, code)
-	})
-}
-
-func TestRunInit_List(t *testing.T) {
-	out := captureStdout(func() {
-		code := runInit([]string{"--list"})
-		assert.Equal(t, 0, code)
-	})
-	assert.Contains(t, out, "Starters")
-	assert.Contains(t, out, "okf")
-	assert.Contains(t, out, "Packs")
-	assert.Contains(t, out, "wordlists")
-}
-
-func TestRunInit_FromMarkdownlint_NoConfig_ExitsTwo(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// No markdownlint config to discover -> initConfigBytes errors ->
-	// writeInitConfig propagates it -> exit 2.
-	captureStderr(func() {
-		code := runInit([]string{"--from-markdownlint"})
-		assert.Equal(t, 2, code)
-	})
-}
-
-func TestRunInit_FromMarkdownlint_Converts(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	require.NoError(t, os.WriteFile(".markdownlint.json", []byte(`{"MD013": false}`), 0o644))
-
-	captureStderr(func() {
-		code := runInit([]string{"--from-markdownlint"})
-		assert.Equal(t, 0, code)
-	})
-	// The converted config ("created from" branch of writeInitConfig) was
-	// written.
-	_, err := os.Stat(filepath.Join(dir, ".mdsmith.yml"))
-	require.NoError(t, err)
-}
-
-// --- applyPacks / writeScaffolds ---
-
-func TestWriteScaffolds_WritesAndSkips(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	files := []pack.File{
-		{Path: filepath.Join(".mdsmith", "wordlists", "a.yaml"), Data: []byte("entries:\n  - x\n")},
-	}
-
-	var buf bytes.Buffer
-	require.NoError(t, writeScaffolds(files, &buf))
-	assert.Contains(t, buf.String(), "created")
-	_, err := os.Stat(files[0].Path)
-	require.NoError(t, err, "file written")
-
-	// A second call skips the now-existing file rather than erroring or
-	// clobbering it.
-	buf.Reset()
-	require.NoError(t, writeScaffolds(files, &buf))
-	assert.Contains(t, buf.String(), "already exists, skipping")
-	got, err := os.ReadFile(files[0].Path)
-	require.NoError(t, err)
-	assert.Equal(t, "entries:\n  - x\n", string(got), "existing file not clobbered")
-}
-
-func TestWriteScaffolds_MkdirError(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// .mdsmith is a real directory, but .mdsmith/wordlists is a regular
-	// file. The parent chain has no symlink, so refuseSymlinkedParents
-	// passes and MkdirAll(.mdsmith/wordlists) then fails because the target
-	// already exists as a file — driving the directory-creation error
-	// branch.
-	require.NoError(t, os.Mkdir(".mdsmith", 0o755))
-	require.NoError(t, os.WriteFile(filepath.Join(".mdsmith", "wordlists"), []byte("x"), 0o644))
-
-	files := []pack.File{{Path: filepath.Join(".mdsmith", "wordlists", "a.yaml"), Data: []byte("entries:\n  - x\n")}}
-	err := writeScaffolds(files, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "creating "+filepath.Join(".mdsmith", "wordlists"))
-}
-
-func TestWriteScaffolds_RefusesSymlink(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	wlDir := filepath.Join(dir, ".mdsmith", "wordlists")
-	require.NoError(t, os.MkdirAll(wlDir, 0o755))
-	// A symlink at the target path must be refused, not followed: writing
-	// through it would land the pack file wherever the link points. Even a
-	// dangling link (its target does not exist) is refused rather than
-	// treated as absent-and-writable.
-	require.NoError(t, os.Symlink(filepath.Join(dir, "escape.yaml"), filepath.Join(wlDir, "a.yaml")))
-
-	files := []pack.File{{Path: filepath.Join(".mdsmith", "wordlists", "a.yaml"), Data: []byte("entries:\n  - x\n")}}
-	err := writeScaffolds(files, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "symlink")
-	assert.Contains(t, err.Error(), "a.yaml")
-	// The link's target was never written.
-	_, statErr := os.Stat(filepath.Join(dir, "escape.yaml"))
-	assert.True(t, os.IsNotExist(statErr), "must not write through the symlink")
-}
-
-func TestStatTarget(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-
-	// Missing path: absent, no error.
-	exists, err := statTarget("missing.yml")
-	require.NoError(t, err)
-	assert.False(t, exists)
-
-	// A regular file exists.
-	require.NoError(t, os.WriteFile("real.yml", []byte("x"), 0o644))
-	exists, err = statTarget("real.yml")
-	require.NoError(t, err)
-	assert.True(t, exists)
-
-	// A symlink is refused, never followed.
-	require.NoError(t, os.Symlink("real.yml", "link.yml"))
-	_, err = statTarget("link.yml")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "symlink")
-
-	// A directory is not a regular file: reported absent (no error) so the
-	// caller's write runs and fails clearly instead of silently skipping.
-	require.NoError(t, os.Mkdir("adir", 0o755))
-	exists, err = statTarget("adir")
-	require.NoError(t, err)
-	assert.False(t, exists, "a directory is not treated as an existing file")
-
-	// A non-ENOENT lstat failure — a path component is a file, so ENOTDIR —
-	// is surfaced as a checking error rather than read as absent.
-	_, err = statTarget(filepath.Join("real.yml", "child"))
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "checking")
-}
-
-func TestApplyPacks_UnknownPack(t *testing.T) {
-	// applyPacks re-checks each name defensively; a bogus name returns an
-	// unknown-pack error even though runInit validates first.
-	err := applyPacks([]string{"bogus"}, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), `unknown pack "bogus"`)
-}
-
-func TestWriteScaffolds_WriteError(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// The target path is itself a directory, so statTarget reports it
-	// absent and os.WriteFile then fails with "is a directory" —
-	// exercising the write-error branch.
-	require.NoError(t, os.MkdirAll(filepath.Join(".mdsmith", "wordlists", "a.yaml"), 0o755))
-
-	files := []pack.File{{Path: filepath.Join(".mdsmith", "wordlists", "a.yaml"), Data: []byte("entries:\n  - x\n")}}
-	err := writeScaffolds(files, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "writing")
-	assert.Contains(t, err.Error(), "a.yaml")
-}
-
-func TestNormalizePackNames(t *testing.T) {
-	assert.Equal(t, []string{"wordlists", "stopwords"},
-		normalizePackNames([]string{"wordlists", " stopwords", "  ", ""}))
-	assert.Empty(t, normalizePackNames([]string{"  ", ""}))
-}
-
-func TestValidatePackPath(t *testing.T) {
-	// Under .mdsmith/ is the contract; an internal ".." that stays inside
-	// is harmless and allowed.
-	require.NoError(t, validatePackPath(filepath.Join(".mdsmith", "wordlists", "a.yaml")))
-	require.NoError(t, validatePackPath(filepath.Join(".mdsmith", "kinds", "..", "a.yaml")))
-
-	// Absolute paths and any escape out of .mdsmith/ are refused.
-	assert.Error(t, validatePackPath(filepath.FromSlash("/etc/passwd")))
-	assert.Error(t, validatePackPath(filepath.Join("..", "evil.yaml")))
-	assert.Error(t, validatePackPath(filepath.Join(".mdsmith", "..", "..", "evil.yaml")))
-	assert.Error(t, validatePackPath(filepath.Join("other", "a.yaml")))
-}
-
-func TestWriteScaffolds_RejectsEscapingPath(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// A pack that returns a path escaping .mdsmith/ is refused before any
-	// write, so nothing lands outside the workspace.
-	files := []pack.File{{Path: filepath.Join("..", "escape.yaml"), Data: []byte("x")}}
-	err := writeScaffolds(files, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "under .mdsmith/")
-	_, statErr := os.Stat(filepath.Join(filepath.Dir(dir), "escape.yaml"))
-	assert.True(t, os.IsNotExist(statErr), "nothing written outside the workspace")
-}
-
-func TestWriteScaffolds_RefusesSymlinkedMdsmithDir(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	// .mdsmith itself is a symlink to another directory; pack writes must
-	// not be redirected through it.
-	target := filepath.Join(dir, "elsewhere")
-	require.NoError(t, os.Mkdir(target, 0o755))
-	require.NoError(t, os.Symlink(target, ".mdsmith"))
-
-	files := []pack.File{{Path: filepath.Join(".mdsmith", "wordlists", "a.yaml"), Data: []byte("x")}}
-	err := writeScaffolds(files, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "symlink")
-	_, statErr := os.Stat(filepath.Join(target, "wordlists", "a.yaml"))
-	assert.True(t, os.IsNotExist(statErr), "nothing written through the symlinked .mdsmith")
-}
-
-func TestWriteScaffolds_RefusesSymlinkedIntermediateDir(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-	require.NoError(t, os.Mkdir(".mdsmith", 0o755))
-	target := filepath.Join(dir, "elsewhere")
-	require.NoError(t, os.Mkdir(target, 0o755))
-	// .mdsmith is a real directory, but an intermediate component
-	// (.mdsmith/wordlists) is a symlink out of tree. The write must still
-	// be refused rather than following it.
-	require.NoError(t, os.Symlink(target, filepath.Join(".mdsmith", "wordlists")))
-
-	files := []pack.File{{Path: filepath.Join(".mdsmith", "wordlists", "a.yaml"), Data: []byte("x")}}
-	err := writeScaffolds(files, io.Discard)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "symlink")
-	_, statErr := os.Stat(filepath.Join(target, "a.yaml"))
-	assert.True(t, os.IsNotExist(statErr), "nothing written through the intermediate symlink")
-}
-
-func TestRefuseSymlinkedParents(t *testing.T) {
-	dir := t.TempDir()
-	t.Chdir(dir)
-
-	// A path whose parent is the cwd (dir == ".") has nothing to check.
-	require.NoError(t, refuseSymlinkedParents("a.yaml"))
-
-	// Parents that don't exist yet are fine — MkdirAll makes them real.
-	require.NoError(t, refuseSymlinkedParents(filepath.Join(".mdsmith", "wordlists", "a.yaml")))
-
-	// A fully real parent chain passes.
-	require.NoError(t, os.MkdirAll(filepath.Join(".mdsmith", "wordlists"), 0o755))
-	require.NoError(t, refuseSymlinkedParents(filepath.Join(".mdsmith", "wordlists", "a.yaml")))
-
-	// A symlinked component is refused.
-	require.NoError(t, os.Symlink(dir, filepath.Join(".mdsmith", "link")))
-	err := refuseSymlinkedParents(filepath.Join(".mdsmith", "link", "a.yaml"))
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "symlink")
-
-	// A non-ENOENT lstat error — a component is a file, so ENOTDIR — is
-	// surfaced as a checking error.
-	require.NoError(t, os.WriteFile(filepath.Join(".mdsmith", "afile"), []byte("x"), 0o644))
-	err = refuseSymlinkedParents(filepath.Join(".mdsmith", "afile", "child", "a.yaml"))
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "checking")
-}
-
 // --- runHelp ---
 
 func TestRunHelp_NoArgs_ExitsZero(t *testing.T) {
@@ -2051,93 +1560,6 @@ func TestDiscoverConfigPath_EmptyNoConfigFallsBack(t *testing.T) {
 
 	got := discoverConfigPath("")
 	assert.Equal(t, filepath.Join(dir, ".mdsmith.yml"), got)
-}
-
-// --- init config generation ---
-
-func TestDefaultConfigBytes(t *testing.T) {
-	data, err := defaultConfigBytes()
-	require.NoError(t, err)
-
-	s := string(data)
-	assert.Contains(t, s, "rules:")
-	assert.Contains(t, s, "front-matter: true")
-	assert.Contains(t, s, "line-length")
-}
-
-func TestInitConfigBytes_EmptyFlagUsesDefaults(t *testing.T) {
-	var buf bytes.Buffer
-	data, source, err := initConfigBytes("", "", &buf)
-	require.NoError(t, err)
-
-	assert.Empty(t, source)
-	assert.Contains(t, string(data), "rules:")
-	assert.Empty(t, buf.String(), "defaults conversion emits no notes")
-}
-
-func TestInitConfigBytes_Starter(t *testing.T) {
-	var buf bytes.Buffer
-	data, source, err := initConfigBytes("", "okf", &buf)
-	require.NoError(t, err)
-
-	assert.Equal(t, "the okf starter", source)
-	assert.Contains(t, string(data), "required-frontmatter")
-	assert.Empty(t, buf.String(), "starter emits no notes")
-}
-
-func TestInitConfigBytes_UnknownStarter(t *testing.T) {
-	var buf bytes.Buffer
-	_, _, err := initConfigBytes("", "nope", &buf)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "unknown starter")
-}
-
-func TestInitConfigBytes_ConvertsNamedConfig(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, ".markdownlint.json")
-	require.NoError(t, os.WriteFile(path,
-		[]byte(`{"MD013": {"line_length": 120}, "MD024": {"siblings_only": true}}`), 0o644))
-
-	var buf bytes.Buffer
-	data, source, err := initConfigBytes(path, "", &buf)
-	require.NoError(t, err)
-
-	assert.Equal(t, path, source)
-	assert.Contains(t, string(data), "max: 120")
-	assert.Contains(t, buf.String(), "siblings_only",
-		"untranslated options are echoed as notes")
-}
-
-func TestConvertedConfigBytes_AutoDiscovers(t *testing.T) {
-	dir := t.TempDir()
-	require.NoError(t, os.WriteFile(filepath.Join(dir, ".markdownlint.yaml"),
-		[]byte("MD041: false\n"), 0o644))
-	t.Chdir(dir)
-
-	var buf bytes.Buffer
-	data, source, err := convertedConfigBytes("auto", &buf)
-	require.NoError(t, err)
-
-	assert.Equal(t, ".markdownlint.yaml", source)
-	assert.Contains(t, string(data), "first-line-heading: false")
-}
-
-func TestConvertedConfigBytes_MissingFile(t *testing.T) {
-	var buf bytes.Buffer
-	_, _, err := convertedConfigBytes(filepath.Join(t.TempDir(), "nope.json"), &buf)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "nope.json")
-}
-
-func TestConvertedConfigBytes_ParseError(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, ".markdownlintrc")
-	require.NoError(t, os.WriteFile(path, []byte(`{"MD013" true}`), 0o644))
-
-	var buf bytes.Buffer
-	_, _, err := convertedConfigBytes(path, &buf)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), ".markdownlintrc")
 }
 
 // failAfterWriter accepts the first n writes, then fails. Drives the
